Modified

+++ datasets/tad_eval.py
@@ -170,9 +170,9 @@ class TADEvaluator(object):
         all_pred = []
         for vid in video_ids:
             this_dets = self.all_pred['nms'][self.all_pred['nms']['video-id'] == vid][['t-start', 't-end', 'score', 'cls']].values
-            b = torch.from_numpy(this_dets[..., 0:2])
-            s = torch.from_numpy(this_dets[..., 2])
-            l = torch.from_numpy(this_dets[..., 3])
+            b = torch.from_numpy(this_dets[..., 0:2]).float()
+            s = torch.from_numpy(this_dets[..., 2]).float()
+            l = torch.from_numpy(this_dets[..., 3]).float()
             b, s, l = batched_nms(
                 b.contiguous(), s.contiguous(), l.contiguous(),
                 0.1,
